Database model, frontend work
I have been searching for a good database design to fit the purposes of this web app.
This is the first draft.
I have been working on adding the data to the database, as well as the plan of study, which will consist of data from several tables. I now need to figure out how to display the data on the websites profile view.